Tried to optimize board extraction but failed. Add train with only portion
