+++ katago/__init__.py
@@ -1,2 +1,2 @@
-from katago.extract_intermediate import extract_features, get_model, get_xy
-from katago import extract_intermediate_optimized
+from katago.extract_intermediate import extract_features, extract_features_batch, get_model, get_xy
+from katago import extract_intermediate
